Add saved listing CSS and animation
