+++ src/scss/_listings.scss
@@ -63,6 +63,7 @@
 }
 
 .listing-container {
+  position: relative;
   font-family: 'Work Sans', sans-serif;
   display: flex;
   flex-direction: column;
@@ -73,6 +74,53 @@
   border-radius: 0.6rem;
   box-shadow: 0 0.25rem 1rem #a4a4a44d;
   overflow: hidden;
+}
+
+.listing-save-container {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 2.4rem;
+  height: 2.4rem;
+  background-color: #292323d8;
+  border-radius: 50%;
+  z-index: 10;
+  right: 1rem;
+  top: 1rem;
+  cursor: pointer;
+}
+
+.heart-icon {
+  font-size: 1.7rem;
+  font-weight: 400;
+  color: #fc5d78;
+  z-index: 10;
+
+  &.saved {
+    font-weight: 900;
+    animation: fill-heart 0.45s linear;
+  }
+}
+
+.heart-dot {
+  position: absolute;
+  width: 0.15rem;
+  height: 0.15rem;
+  border-radius: 50%;
+  top: 50%;
+  left: 50%;
+  opacity: 1;
+  transform: translate(-50%, -50%);
+
+  &::before {
+    content: "";
+    position: absolute;
+  }
+ 
+  &.saved {
+    animation: move-dot 0.8s ease-out forwards 0.3s;
+  }
 }
 
 .no-images-text {
@@ -99,8 +147,12 @@
 .listing-image {
   width: 100%;
   height: calc(100% - 2rem);
+
+  min-height: 10rem;
+
   object-fit: cover;
   margin: 0 auto;
+  background-color: white;
 }
 
 .img-arrow {
@@ -144,26 +196,40 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  gap: 0.25rem;
+  gap: 0.25rem 0.45rem;
   width: 100%;
   height: 2rem;
   background-color: eee;
 }
 
 .listing-image-circle {
-  font-size: 2rem;
   color: #0a1064;
-  width: 1rem;
-  transition: 300ms ease all;
-  -webkit-transition: 300ms ease all;
+  width: 0.75rem;
+  height: 0.75rem;
+  transition: 200ms ease all;
+  -webkit-transition: 200ms ease all;
+  border: 2px solid #0a1064;
+  border-radius: 50%;
   cursor: pointer;
+  user-select: none;
+  -webkit-user-select: none;
+
+  &:hover {
+    background-color: #0a1064;
+    transition: 200ms ease all;
+    -webkit-transition: 200ms ease all;
+  }
 }
 
 .listing-image-circle-current {
   color: #1a75ff;
-  font-size: 2.75rem;
-  transition: 300ms ease all;
-  -webkit-transition: 300ms ease all;
+  background-color: #1a75ff;
+  width: 1rem;
+  height: 1rem;
+  border-radius: 50%;
+  font-size: 1rem;
+  transition: 200ms ease all;
+  -webkit-transition: 200ms ease all;
 }
 
 .listing-details-container {
@@ -232,16 +298,13 @@
 }
 
 .listing-link-container {
+  position: relative;
+  display: inline-block;
   font-weight: 600;
   text-align: center;
   text-transform: none;
   padding-top: 0.25rem;
   margin: auto auto 0 auto;
-}
-
-.listing-link-container {
-  position: relative;
-  display: inline-block;
 
   &:hover .listing-link {
     width: 100%;
@@ -251,16 +314,18 @@
 .listing-link {
   position: absolute;
   display: inline-block;
+  color: #9590d1;
   width: 0;
   white-space: pre;
   overflow: hidden;
   transition: width 500ms;
-  border-bottom: 1px solid #9590d1;
+  text-decoration: underline;
 }
 
 .listing-link-default {
   color: #222427;
   text-decoration: none;
+  padding: 0.5rem 0;
 
   &:hover {
     color: #222427;
@@ -271,6 +336,7 @@
 .listing-link-hover {
   color: #6f6ca8;
   text-decoration: none;
+  padding: 0.5rem 0;
 
   &:hover {
     color: #6f6ca8;
